general/rb: make decimal_time prettier
One would think skipping a require and only creating a single time
object from which the number of seconds passed in the day are calculated
would speed things up (or slow them down because datetime is so good or
something). But no. It literally doesn't make a difference even if the
script is run 1000 times. Trust me, I tried.